long! definitely long!
Former-commit-id: d58aa6ae4e8b39201107cb0414df96881ff3cbe6

+++ palladian-commons/src/main/java/ws/palladian/helper/math/MathHelper.java
@@ -965,14 +965,14 @@ public final class MathHelper {
 
     /**
      * Map two natural numbers (non-negative!) to a third natural number. N x N => N.
-     * f(a,b) = c where there are now two settings for a and b that produce the same c.
+     * f(a,b) = c where there are no two settings for a and b that produce the same c.
      * @see https://en.wikipedia.org/wiki/Pairing_function
      * @param a The first number.
      * @param b The second number.
      * @return The target number.
      */
-    public static int cantorize(int a, int b) {
-        return ((a + b) * (a + b + 1) / 2) + b;
+    public static long cantorize(int a, int b) {
+        return (((long)a + (long)b) * ((long)a + (long)b + 1) / 2) + (long)b;
     }
 
     /**
